rename component file to include drilldown
