+++ packages/client/hmi-client/src/components/workflow/ops/optimize-ciemss/mod.ts
@@ -1,6 +1,6 @@
 import { OptimizeCiemssOperation as operation } from './optimize-ciemss-operation';
 import node from './tera-optimize-ciemss-node.vue';
-import drilldown from './tera-optimize-ciemss.vue';
+import drilldown from './tera-optimize-ciemss-drilldown.vue';
 
 const name = operation.name;
